Change ClassPredictionError Chart aesthetics
There was a lack of space between the tallest bar and the upper
edge of the chart.  I increased the y-lim relative to the tallest
bar.

Additionally, hard-coded plot attributes were removed so that
they will be user-defined instead.  For instance, the grid parameter
was removed and is at it default setting of True.