This commit will: Update code to reflect reviewer suggestions
The overarching aim of this commit is to tackle the addition of different colors to positive and negative coefs.
Ths commit specifically,is necessary in order to address concerns brought up in the previous code review.

The concerns being the following:

We would like to allow the user to set the color of the bars.
However, the color argument was implicitly set (ie available for use but user didnt know this)
    * Added a color argument directly to __init__ of FeatureImportances Class

The following scenario was proposed.
If the users pass color as a string(e.g. A single color for feature importances bar)
What happens?  What happens if there are positive and negative classes?

    * In the original commit, strings were inherently handled thus no code was changed to address this.
    * It gets really interesting when addressing what happens if there are positive and negative classes.
        * Firstly, I determined if there were indeed both positive and negative values.
          This information helped to define the number of colors a graph can support.
          For example if both pos/neg values are present then the graph can support 2 colors.
          Based on how many colors a graph could support I then was able to assign different colors
          for positive and negative coefficients or a single color for all values.