If added this commit will add a test stub for test_importance
This is my attempt to write test stubs for test_importance.py
I struggle with writing test because I am not sure what I should
be testing.

For this stub,  I hope to catch the error when a user specifies
a single color for a graph although there are both positive and
negative coef present(ie the graph supports 2 colors).
    * It is assumed that the user selects a single color with
      with hopes that both negative and positive will have the
      same color.  In practice the user can overcome this error
      by specifying a string of two the same color character.