Fix tSNE bugs injected by numpy 1.13  (#333)
Fixes #323 that was reported on StackOverflow: a change in the way matrix comparisons are handled in Numpy 1.13 caused our visualizer to break. 

To fix this, I changed vector existence comparisons to `is None` to explicitly query them, 
though this will not protect against empty vectors. 

I also added several new test cases and image comparisons, but to do that I needed to add a `random_state` to the visualizer to ensure that all stochastic components behaved as expected. Python `set` objects are randomly hashed, so I changed to using `np.unique`. This caused us to have to remove the class filtering feature of tSNE, but it is more robust. 

Another note: the tests also use PY3 vs PY2 tolerances to improve Travis testing.