Removes font size hardcode from confusion matrix (#343)
Fixes #301

The font size was hardcoded to 8, removed this to ensure that it would be easier to read in the future. Added a fontsize argument the visualizer in order for the user to manually specify fontsize, otherwise, the default is taken from mpl.rcParams. 

Refactored the visualizer to meet pyflakes checking and to make it a bit simpler. Deprecated passing options to fit, and instead hoisted them to the constructor. Also added more visual test cases. Percent mode is now no longer the default.