Add import statements to code blocks for classification visualizers
Add import statements to code blocks so that scikit-learn, yellowbrick, and other neccessary modules are imported. These import statements are included on the index page for classification visualizers; however, if someone previously went directly to one of the individual pages and tried to run the code on their own, they would get an error because the necessary modules were not included. These changes will make it easier for new users to get started using Yellowbrick.