Updates ClassificationReport to use pcolormesh (#348)
Fixes #267

Updates the classification report to use the pcolormesh instead of
imshow so that the grid lines are on the outside rather than through
the figure numbers. This also makes it more robust to non-square CR
shapes and odd grids.

Additionally adds a quick method to ConfusionMatrix and tests for 
both the confusion matrix and classification report quick methods.