Make sklearn classification dataset for Class Prediction Error Tests (#355)
In several places throughout the code, you will find that some test data is hardcoded.
This is the X/y data that is used by our models.  I originally used this data to test
my visualizer.  However, I soon found out that the data isn't suitable for all
applications.  In my case, using this data produced completely segmented clusters and this
isn't useful when you are trying to look at prediction error.

My solution is simple. I use the `sklearn.make_classification` function with a random state set
This allows for more complex data sets in which I am able to specify the number of classes
and how many are informative(plus more parameters).  I think the most important thing
is the setting the random state, this allows for completely reproducible images.