Fix documentation: Add import statements to documentation for individual visualizers (#350)
* Documentation updates

* Documentation updates

* Deleted duplicate paragraphs on the documentation's 'About' page.

* Add import statements to code blocks for classification visualizers

Add import statements to code blocks so that scikit-learn, yellowbrick, and other neccessary modules are imported. These import statements are included on the index page for classification visualizers; however, if someone previously went directly to one of the individual pages and tried to run the code on their own, they would get an error because the necessary modules were not included. These changes will make it easier for new users to get started using Yellowbrick.

* Add import statements to code blocks for clustering visualizers

* Add import statements to code blocks for regression visualizers

Add import statements to code blocks so that scikit-lear, yellowbrick, and other necessary modules are imported. These import statements are included on the index page for the regression visualizers; however, they were not originally added to the individual pages in the documentation.

* Add import statements to code blocks for text visualizers

* Add import statements to code blocks for features visualizers