Adds VisualizerGrid (#176)
Adds VisualizerGrid - a basic implementation of the `MultipleVisualizer` base class that creates axes for each visualizer using plt.subplots, laying the visualizers out as a grid. Users can specify the number of rows or columns in the layout. When fit, predict, poof, etc. are called, the underlying visualizer methods are called in order. 

TODO for multiple visualizers:

- Fix tests 
- Make a MultipleVisualizer a Sequence 
- Implement FlipBook visualizer