Modifies Scoring Mechanism for Regressor Visualizers
This PR is meant to start changing the score visualizer methodology.
Besides simply being the entry point for drawing, the score methods
should also return the scoring function of the underlying model. This
commit repairs the regressor models and also updates them to include
the R2 value in the plot itself with the legend.

Still to do are the classifier visualizers.

This commit also updates the PredictionError and ResidualPlot
integration tests to use better random data and the PyTest method
as well as includes support for pandas integration testing. I've also
updated the documentation with the new visualizers.

Fixes #213