Updates the discrimination threshold visualizer and type checking tests. (#364)
* Updated ThreshViz to DiscriminationThreshold
* Refactor of tests for type checking utilities

Updates the ThreshViz class to DiscriminationThreshold and implements a
few more discrimination features such as F1 score, maximizing arguments
and annotations. Documentation and tests are also updated to bring this
visualizer into prominence and give more explanation.

Added `is_probabilistic` type checker.

A major win here was to convert the type checking tests to pytest.
This has the advantage of moving our tests to pytest parametrize so that
we can get better coverage of test cases. Tests in this file have jumped
from 24 tests to 165 tests! Feeling much better about these utilities as
a result.

Fixes #84