Model Selection Package
Initial proposal for a model selection package that mirrors sklearn
model selection. Since most of what YB does is about model selection I
was unsure if this is the correct name. If it is, then grid search
should be placed under this library.

Alternatively we could call this package cross_validation and place CV
visualizers such as the one @pdamodaran is working on and the learning
curve visualizer in this package and keep this separate from the grid
search module. However this might be confusing because of sklearns
deprecation of the cross_validation package.