Model Selection Package (#377)
* Added Model Selection Package with LearningCurve and ValidationCurve 

Initial proposal for a model selection package that mirrors sklearn
model selection. Since most of what YB does is about model selection I
was unsure if this is the correct name. If it is, then grid search
should be placed under this library.

* Updates to LearningCurve

Added image comparison tests for the LearningCurve visualizer and
created a documentation page for it. Moved LearningCurve to the model
selection package.

Additionally updates test dataset loader to load pandas dataframes if
available and generalizes the fixture lookup and download process.

Fixes #116

* Added a ValidationCurve visualizer

Validation curves are a sibling to learning curves and visualize the
model performance with respect to a single hyperparameter. Often
validation curves can be used to demonstrate the bias/variance trade-off
and can be used to steer grid search methodologies.

Fixes #378