Added skip if Windows for image comparisons
If image comparisons were not close on windows, added a skip decorator
to skip the test on Windows. This is an intermediate solution to getting
yb tests working on Windows, but would like to ensure that we provide
more robust support for multiple operating systems in the future.

Note: also added some os.path.join test fixes instead of relying on
hardcoded paths.