Added AppVeyor configuration and ImageComparison encapsulation (#386)
* Added AppVeyor configuration
* Skip tests on Windows 32 bit architectures.
* use NLTK module downloader
* Added xfail if Windows for image comparisons

AppVeyor is a CI tool for Windows builds. Added CI for Python 2.7 and
3.6 on Windows as well as Miniconda 4.4.10.

If image comparisons were not close on windows, added xfail decorator
to complete the tests on Windows. This is an intermediate solution to getting
yb tests working on Windows, but would like to ensure that we provide
more robust support for multiple operating systems in the future.

Note: also added some os.path.join test fixes instead of relying on
hardcoded paths.

* Encapsulation of ImageComparison for Windows tests

This is sort of a large commit, I wrapped all of the image comparison
logic for assert_images_similar into a single class, ImageComparison.
The ImageComparison object is callable and saves the state of a single
assertion statement (removing it from the TestCase so we can refactor to
using all pytest instead of unittest). During this encapsulation I also
added a lot of os.path handling to ensure that tests worked (hopefully)
on Windows as well as on Posix, continuing the integration with
AppVeyor.

ImageComparison now also raises an assertion error when nothing has been
drawn on the axes object supplied (even if a title or axis labels have
been added). This is further safety for some of our tests that just had
empty images in them.

Updated the ImageComparisonFailure exception to use our own internal
exception instead of matplotlibs and made the exceptions a lot more
readable.

Part of the refactor included a test_meta.py tests, which actually tests
the ImageComparison class and some of our other testing utilities like
the RandomVisualizer.

Updated ROCAUC to raise an exception on no curves instead of drawing an
empty image. Updated test_scatter tests to use the correct
ImageComparisonFailure exception class.