Manifold Learning Visualization
Manifold learning includes algorithms for random projection of high
dimensional data into a 2 or 3 dimensional space for visualization.
Unlike PCA or SVD, manifold learning does a better job of embedding
points that have non-linear relationships, making them useful for
detecting clusters in data and therefore separability for learning
algorithms.

Because there are a variety of manifold algorithms, this visualizer
groups them all into a single visual framework, allowing specificication
of which model to use by supplying an estimator or a string that refers
to the algorithm in question.

Implements #398

* this first commit shows the files required for starting a visualizer.