Feature: If applied this commit will include support metric in the classification reports
This commit will fix issue #359 by including the support metric in the classification
reports. Prior to this change, the code assumed we did not want to display the
support metric, so it was not included in the SCORES_KEYS, and when it was calculated
with the precision_recall_fscore_support function, the support metric was not
included in the rest of the code.

Now, the classification report displays the support metric for all classes. The
support is *displayed* as the percent of the total number of records using a heat
map color system but it is *labeled* with the actual support number.

Resolves: Issue #359