Fix: If applied, this commit will include testing for the support metric on the classification reports
This commit will continue fixing issue #359 by finalizing the display of the support metric
on the classification reports as well as changing the classification report testing to ensure
that it includes the support metric. Prior to this change, the support metric was only
available for two classes and would break if there were more than two classes to compare. It
would also fail most of the unit tests because the support metric was not included in the
unit testing.

Now, it is capable of working with more than two classes and the unit tests account for the
new support metric feature.

Resolves: Issue #359