Docs: If applied, this commit will include documentation about the support metric in the classification reports.
This commit will include documentation about the support metric to reflect the fact that the support metric is now supported for the classification report. Prior to this change, the documentation stated that the support metric is not included int he report.

Now, the documentation includes information about the support metric.

Resolves: Issue #359