Added continuous and discrete colors
Created a lightweight method to detect continuous or discrete targets
then embedded the color selection methodology into the class. When no
target is specified, a single color is used. When a discrete target is
detected (<10 unique values) or specified, then resolve colors is used
and a legend is added to the plot. When a continuous target is detected
or specified, then a colormap is used and a colorbar is added to the
plot.