Manifold Learning Visualization (#399)
Manifold learning includes algorithms for random projection of high
dimensional data into a 2 or 3-dimensional space for visualization.
Unlike PCA or SVD, manifold learning does a better job of embedding
points that have non-linear relationships, making them useful for
detecting clusters in data and therefore separability for learning
algorithms.

Because there are a variety of manifold algorithms, this visualizer
groups them all into a single visual framework, allowing specification
of which model to use by supplying an estimator or a string that refers
to the algorithm in question.

* Added continuous and discrete colors

Created a lightweight method to detect continuous or discrete targets
then embedded the color selection methodology into the class. When no
target is specified, a single color is used. When a discrete target is
detected (<10 unique values) or specified, then resolve colors is used
and a legend is added to the plot. When a continuous target is detected
or specified, then a colormap is used and a color bar is added to the
plot.