Add random sampling and improve performance of Parallel Coordinates (#420)
To improve the drawing time of the Parallel Coordinates visualization, this PR 
implements two new functionalities:

1. allow random subsampling with shuffling to reduce the number of lines drawn 
2. instead of plotting individual traces, consolidate traces into class and plot each class 

The visualizer fit state is adapted to compute the number of samples in the drawing 
and a random state parameter is added to test the static sampling/shuffling state. 

Note that this PR is complete, though some tests are failing (to be fixed in the 
next commit), but it does change the style of the output visualizer. See #447. 

Additionally, adds the is_series type check utility to test if y is a pandas.Series 
object.