feat: If applied, this commit will make 'support' optional for classification report
This change is being made because the user had no control over when and how to
display support on the classification report visualizer.

It's being made to allow the user to select if the support will be showcased
as a raw count of a percentage.

This change is also being made because we want the new feature to be backward-compatible

Expands PR #413
It provides