Add CVScores visualizer to model selection package (#444)
Adds a CVScores visualizer to visualize the cross-validation process. The visualizer creates a bar graph of the scores for each split and annotates across each split the mean or final score computation. This package also includes a quick method, cv_scores. 

NOTE: This visualizer is currently undocumented and untested. See #464 to make this visualizer documented.

Closes #374