Deprecate ScatterVisualizer to contrib
To mantain backward compatibility, the ScatterVisualizer has been moved
to contrib, so that the new ScatterPlotMixin (which borrows heavily from
this visualizer) does not break any user code.

Moved the ScatterVisualizer to contrib, updated documentation and
tests. No changes were made to the visualizer, but it can be updated to
use the new ScatterPlotMixin.

Note that documentation now lists "direct data visualization" as only
the JointPlotVisualizer, but we should add some of the single feature
visualizations to direct data visualization as well.

Fixes #440