Implements histogram alongside ResidualsPlot (#480)
* Implements histogram alongside Residuals Plot

Adds a hist=True option to the ResidualsPlot visualizer, which plots the
histogram of the residuals as a barh on a vertical axes with sharey=True
to the right of the scatter plot axes. This allows the user an easier
method to identify the distribution of the errors (e.g. more positive or
more negative).

Note that matplotlib 2.0.2 or greater is required for this method.

Fixes #264