Add tests to test ignore_case and usage of a generator as a corpus input
We first cleaned up the primary test according to review by Ben.  This
included switching to immutable arrays and adding np.flip() to reverse
the target_word_ list

We generate two additional tests.  The first is to test if the corpus can
be inputed as a generator.  The second is to test the plot with ignore_case
equal to True