Feature importances fix (#490)
Adds a check to feature importances for coefficients that take the form of multidimensional arrays, and if so, takes the mean of the arrays by column to flatten them to enable the feature importances visualizer to successfully plot the importances as a bar chart. Includes tests for the case. 

Fixes #485