Fixes Manifold fit_transform bug (#505)
Some manifolds do not provide a transform() method since the data
transformation process cannot be decoupled from the fit() process. In
this case, the user is required to call fit_transform() to get the
transformed data. Unfortunately, the Manifold visualizer implemented the
TransformerMixn and fit() and transform() methods, so when calling the
default fit_transform() method an exception was raised.

Because the fit() method already called fit_transform() - the fix was
simple. Manifold now overrides fit_transform() to do the work that fit()
used to do, and fit() simply calls fit_transform(). In addition, I've
captured AttributeError exceptions in transform() and reraised them with
a message that indicates the user should use fit_transform() instead.

Tests have been added to ensure fit() calls fit_transform(), that
fit_transform() works as expected, and that transform() raises the
correct exception.

Fixes #504