Add Examples Notebook of Regression Visualizers (#470)
Initially, I had intended to update the original examples.ipynb, which contains a brief description and example of the current visualizers available in Yellowbrick. However, that notebook was already quite long, even before adding our newest visualizers. So I decided to create a new series of notebooks that focus on only one of our main categories of visualizers.

This notebook contains examples of Yellowbrick's regression visualizers. To make it easier to review, I extracted the exact code and description text from the original examples.ipynb and tried to make only minimal changes. Some of the more notable changes are:

* Uses the new yellowbrick.datsets module to load the data
* Adds text from our documentation to expands on the Visualizer descriptions
* Adds an additional example of the Residuals Plot to show how to turn off the optional histogram
* Replaces the code used to create the list of alphas for the AlphaSelection Visualizer with the code used in the documentation. I think this plot is preferable since it clearly shows the error dropping as it approaches the optimal value of alpha and then rising again after it passes it.
* Doesn't include the ManualAlphaSelection Visualizer. The example in the original is entirely commented out and when returns an error when run, so I've left it out for now until I've had a chance to find what is causing the problem.