Creates target module and moves ClassBalance to it (#511)
The target module allows independent analysis to be conducted on the
dependent variable in supervised machine learning. For example, analysis
of label encoding, binarization, or thresholding mechanisms, to simpler
analyses like class balance or feature correlation. The base class,
TargetVisualizer, implements a fit method that only accepts a vector y,
similar to the LabelEncoder transformer in scikit-learn.

By moving the ClassBalance visualizer to this package, it no longer
requires a Model (and is not a score visualizer) so the class balance
analysis can be done a priori without requiring a model. This seems more
in line with many workflows that create train/test splits then would
like to know the class balance of those splits before moving on to
evaluation methods.

ClassBalance has also been extended to provide two modes. In "balance"
mode, the default mode, the bar graph is simply drawn for each class. In
"compare" mode, the train and test splits for each class are drawn side
by side.

Fixes #334