ClassBalance backwards compatibility warnings (#513)
Added a warning to the documentation about the change to the
ClassBalance signature in version 0.9. This can be removed after version
1.0 or later, but should stick around for a few months at least.

Added a check to test the number of dimensions of the first argument to
fit() in ClassBalance. If ndims == 2, I presume this to be following the
fit(X, y) pattern, and raise an exception notifying the user of the
signature change.