Adds undocumented BalancedBinningReference visualizer (#518)
The BalancedBinningReference is created to give references for binning samples into balanced classes. When the target is continuous, it is often not possible to predict using regression; transforming the problem into a classification by binning scores is a good solution. Determining thresholds is difficult, however, as unbalanced samples in each bin might be harmful to the performance of the model.

To solve this problem, I decided to create a visualizer to give a reference for binning balanced samples into each bin.

Fixes #312