Added the Dataset and Corpus data structures
This is a significant change that perhaps we should save for the 1.0
release of Yellowbrick. Datasets, instead of being simple numpy arrays,
are now complex objects that provide a lot of functionality similar to
the sklearn Bunch object; e.g. access to README and meta information,
the ability to transform data into arrays or data frames and more.

This represents a complete overhaul of how datasets in Yellowbrick are
managed and will require updates to the documentation and tests.