Corrects the color of the test residuals in the ResidualPlot quick method
There is a coding error in residualPlot quickmethod.  The test and train
residuals were both being plotted on the scatter plot as blue.
I found that this is because both `train_color` and `test_color` parameters
were being set to train_color which in this case is blue.  This commit
changes the `test_color` parameter equal to test_color