Fixes binary ROCAUC with decision function only models bug (#533)
This PR fixes a bug that plagued ROC/AUC when visualizing binary classification. In this specific case when the model has only a decision_function and not predict_proba, the shape of the returned predictions is 1 dimensional, and not the expected two dimensions. This fix detects this scenario and can draw a binary ROCAUC curve if the micro, macro, and per class options are set to False.