Add notebook with clustering visualizer examples
The Jupyter notebook contains examples of Yellowbrick's clustering
visualizers. It includes the KElbowVisualizer and SilhouetteVisualizer
examples used in the original examples.ipynb with a few minor changes:

- Adds text from Yellowbrick's documentation to expand on the Visualizer
descriptions
- Adds an additional KElbowVisualizer chart to demonstrate how to select a
different scoring metric and to hide the time to fit the model.
- Adds an additional SilhouetteVisualizer plot using the optimal number
of clusters so that readers can compare it to the plot using a lower
than optimal number of k.

Progress on issue #470