Update faq.rst
So I read through this and am ready to submit my review.  I think it's great and will be helpful.  Awesome job particularly on the code blocks and the salience of the questions selected (they will help users customize and save their work which is a tremendous way to engage them and keep them using YB!).  I did change some of the wording (nothing was wrong but just tried to do what I do) and I did test the sample code blocks on my demo Jupyter notebook and it all worked.  I will send this in as a pr as a new branch so that you can take a final look.  For some reason when I clicked on the doc links they referred back to the FAQ page so maybe check that out but I think it's great.