Adds Intercluster Distance Map visualizer
Intercluster distance maps display an embedding of the cluster centers in
2 dimensions with the distance to other centers preserved. E.g. the closer
to centers are in the visualization, the closer they are in the original
feature space. The clusters are sized according to a scoring metric. By
default, they are sized by population, e.g. the number of instances that
belong to each center. This gives a sense of the relative importance of
clusters. Note however, that because two clusters overlap in the 2D space,
it does not imply that they overlap in the original feature space.

Fixes #571