Define boundaries and points in a single loop and bring scatterplot to foreground
We wanted to input our text as a list of documents that contain as list of words.
The problem I faced was that the dispersion plot was only as long as the longest document.
The hope was to treat the corpus documents as one continuous text.
This was accomplished by using an index variable that keeps count
of each word and the count increments even between documents.  The index
value is stored at the end of each document which is later used to draw
the document boundaries

Use the 'zorder' parameter in scatterplot to push points in front of vertical lines.