Adds Precision-Recall Curves Visualizer (#599)
* Adds Precision-Recall Curves Visualizer

The Precision-Recall curves visualizer plots the precision against the
recall of a classifier for different thresholds produced by a decision
function. This visualizer works both with the binary and multiclass
cases, in the latter case it uses a OneVsRestClassifier strategy to
compose a precision-recall matrix.

The visualizer supports multiple visual modes, including area mode,
iso-f1 curves, micro averages and per-class curves (for the multiclass
case). This means that the PRCurves visualizer has a lot of different
visual expressions depending on the data and what the user is trying to
achieve.

Fixes #596