TSNE Visualizer global bug fixes
Hi Rebecca,

Merry XMass and Happy New Year.

Your message is double:

1- about my bug, the problem is very difficult to detect as each time you visualize the TSNE visualizer, you obtain a different figure due to the random character of the mapping.
The bug is not observed in the attached notebook. It affects the resorting of the colors when you use the np.unique() method. My fix avoids this resorting and gives a consistent match between labels, colors and real data

2- the notebook you have done shows a bug when the both colors and labels are passed to the constructor and the fit() method is used only with the docs. 
I think that it is due to a confusion in the code between the classes and the labels : if the labels are passed in the constructor and if they are passed in the fit() method, they were not used the same way and several issues affected the original code.
I have done a fix that I attach to this email as a notebook. 
The TSNEVisualizer_fixed class includes the both bug fixes : the first one for the labels/color resorting problem discussed before and the second for the crash we discuss in this second bullet item.

If you agree with these both fixes, you can use the code in the notebook after you have removed the ## DEBUG code bricks :)

Best regards

Jerome