Extend FeatureImportances docs/tests for stacked importances (#678)
Wraps up the enhancement to the FeatureImportances visualizer which
added a stacked bar chart optional parameter in the case of
multi-dimensional importances. Updated the documentation to reflect the
stack=True situation as well as issued a warning if stack is False but
should be True. Updated tests for better coverage.

Closes #531