Fix errors related to release of pytest 4.2
The release of pytest 4.2 broke some of our custom pytest-spec related
code (e.g. how the test report is printed). This PR attempts to solve
some of the issues, but I've also opened an issue with pytest-spec to
see if we can resolve some of the warnings and add the parametrize ids
back into the code base.