Adds plot directives to quickstart (#719)
This PR was primarily intended to add plot directives to the quickstart
guide, ensuring that the images in the tutorial were always up to date
with the library. The quickstart guide separates the code blocks from
the plot directives so that the code is a linear narrative rather than
the verbosity required for each independent code block. This duplicates
the code a little bit but makes it more readable.

To ensure that the code is correct I also created a notebook
examples/walkthrough.ipynb with the code from the quickstart.

Along the way there were some bugs with jointplot and rankd that I also
fixed. The JointPlotVisualizer needs some more work but it is now
stable.

Additionally, in poof() if the visualizer didn't have an axes object it
would exit silently. This made it hard to find bugs. Now instead of
exiting we simply issue a warning and carry on.