Implement plot directive for discrimination Threshold
1.  Note that two images are produced by the old image generation script.
while only one image is used in the documentation

One uses spam dataset and the other churn dataset.

Ben asked that I include both images in the new documentation since they both look great.
However, set :nofigs: True and :include_source: False in the plot directive for the churn image
Ben suggests that I should post a comment in the PR that the inclusion of the second image is up for review
Ben also asked that I include both images in PR

2.  The churn data is missing.  There is no actual file.  I messaged Kristen to see if she is missing it also.