Tips and conventions for contributors and reviewers (#731)
This PR is intended to refresh the contributor's docs a bit, adding in a few more of our tips and conventions, specifically around things like installing the library in editable mode, merging in PRs, and using feature branches. It follows up on #689 and a few of the conversations that have come up amongst the maintainers recently.