Revert to "Previous Visual Studio 2015" as temporary workaround for miniconda CI (#743)
This fix addresses a CI bug that was causing persistent build failures for miniconda in Appveyor. The primary solution for now is to use the previous image of visual studio as temporary work around, but this PR also bumps our miniconda version and six, adds conda to path, and removes tony from Appveyor automated notifications.