Bug fix for PR curve multiclasses labels (#688)
When the `classes` argument is provided to `PrecisionRecallCurve` there is a mismatch between the contents of `y` (usually numerically encoded) and the text labels that cause `label_binarize` to fail. The fix is to create an intermediate data structure to hold the mapping while binarizing and for labeling purposes.

Fixes #601.