Changed labels to y vector,passed labels as list
In this pr,I have changed the docs and labels to X and y vector in the fit method which created confusion in tsne documentation.
Also,I have passed labels as a list to the TSNEVisualizer(labels=labels). I am attaching the image poofed after passing labels as a list.
fixes #684.