Update about page with affiliatons
In response to recent queries about our About page, I've updated it for
clarification -- particularly the part about affiliations. Although this
page will probably change a bit when we ratify the new governance
document, I think this is a start in the right direction. It also takes
care of some of the action items regarding the clarification of our
relationship with DDL.