Add elbow detection using the "kneedle" method to Elbow Visualizer (#813)
With permission, we have ported the github.com/arvkevi/kneed library into our Yellowbrick utilities for knee/elbow point detection in curves. Our first use of this package is to identify the optimal `k` in the `KElbowVisualizer`, annotating it with a dotted line if found. Note that this commit does not contain tests, though it does update the documentation. Tests to follow shortly.