Add try/except to import calinski_harabasz_score
In scikit-learn 0.21.2 the calinski_harabaz_score method was deprecated
in favor of calinski_harabasz_score in order to fix a typo in the
original name.

When we implemented this change in PR #887, it unintentionally injected
an ImportError for our current set of dependencies since
calinski_harabasz_score does not exist in scikit-learn 0.20.0.
Therefore, this PR fixes Issue #902 by wrapping the import statement
in try/except.